refactor(points): rm empty implementation from superclass
